+++ src/BaseLinter.php
@@ -26,7 +26,7 @@ class BaseLinter
         $this->codeLines = preg_split('/\r\n|\r|\n/', $code);
     }
 
-    public static function appliesToPath(string $path): bool
+    public static function appliesToPath(string $path, array $configPaths): bool
     {
         return true;
     }