+++ src/Linters/Concerns/LintsMigrations.php
@@ -4,7 +4,7 @@ namespace Tighten\TLint\Linters\Concerns;
 
 trait LintsMigrations
 {
-    public static function appliesToPath(string $path): bool
+    public static function appliesToPath(string $path, array $configPaths): bool
     {
         return strpos($path, 'migrations') !== false;
     }